+++ common/common_lib/src/main/java/ru/vm5277/common/RTOSLibs.java
@@ -24,7 +24,13 @@ public enum RTOSLibs {
 	INCTANCEOF("j8b/instanceof.asm"),
 	CLEAR_FIELDS("j8b/clear_fields.asm"),
 	DISPATCHER("core/dispatcher.asm"),
-	INVOKE_METHOD("j8b/invoke_method.asm");
+	INVOKE_METHOD("j8b/invoke_method.asm"),
+	MATH_MUL8("math/mul8.asm"),
+	MATH_MUL16("math/mul16.asm"),
+	MATH_MUL32("math/mul32.asm"),
+	MATH_DIV8("math/div8.asm"),
+	MATH_DIV16("math/div16.asm"),
+	MATH_DIV32("math/div32.asm");
 	
 	private	final	String	path;
 	private			boolean	required;